requirements.txt
Flask==3.0.3
flasgger==0.9.7.1
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
psycopg2-binary==2.9.1
fastapi==0.70.0